Moved convert_param_to_int to a util
